pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "numpy>=2.2.6",
    "pandas>=2.2.3",
    "pillow>=11.2.1",
    "streamlit>=1.45.1",
    "tensorflow>=2.14.0",
    "tensorflow-hub>=0.16.1",
]
